Revert "Extracted UI class + refactored GameController"
This reverts commit ea213554279b405cc58791b754e910ca5537a487.
